+++ spec/lib/proxy_spec.rb
@@ -118,6 +118,29 @@ shared_examples_for 'a cache' do
     end
   end
 
+  context 'cache_whitelist GET requests' do
+    before do
+      Billy.config.whitelist = [http.host]
+      Billy.config.cache_whitelist = [http.host]
+    end
+
+    it 'should be cached' do
+      assert_cached_url
+    end
+
+    context 'with ports' do
+      before do
+        rack_app_url = URI(http.url_prefix)
+        Billy.config.whitelist = ["#{rack_app_url.host}:#{rack_app_url.port + 1}"]
+        Billy.config.cache_whitelist = Billy.config.whitelist
+      end
+
+      it 'should be cached' do
+        assert_cached_url
+      end
+    end
+  end
+
   context 'ignore_params GET requests' do
     before do
       Billy.config.ignore_params = ['/analytics']
@@ -232,7 +255,7 @@ shared_examples_for 'a cache' do
 
         it 'should raise error when disabled' do
           # TODO: Suppress stderr output: https://gist.github.com/adamstegman/926858
-          expect { http.get('/foo') }.to raise_error(Faraday::Error::ConnectionFailed, 'end of file reached')
+          expect { http.get('/foo') }.to raise_error(Faraday::ConnectionFailed, 'end of file reached')
         end
       end
 
@@ -261,7 +284,7 @@ shared_examples_for 'a cache' do
         end
 
         it 'should raise error for non-successful responses when :error' do
-          expect { http_error.get('/foo') }.to raise_error(Faraday::Error::ConnectionFailed)
+          expect { http_error.get('/foo') }.to raise_error(Faraday::ConnectionFailed)
         end
       end
     end